# contrato/openapi.json
{
  "components": {
    "schemas": {
...
      "BriefEntrada": {
        "additionalProperties": false,
        "description": "Lo que hace falta para dar de alta una `Obra`.\n\nSolo `titulo` y `premisa` son obligatorios, que es lo que dice la ficha de\n`Obra`. `persona` y `tiempo_verbal` son de `GuiaDeEstilo` y entran aqui\nporque sin ellos el Escaletador no puede empezar; el resto de la guia llega\ncon su propia feature.",
        "properties": {
          "extension_objetivo": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Objetivo"
          },
          "genero": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genero"
          },
          "persona": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonaNarrativa"
              },
              {
                "type": "null"
              }
            ]
          },
          "premisa": {
            "minLength": 1,
            "title": "Premisa",
            "type": "string"
          },
          "subgenero": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subgenero"
          },
          "tiempo_verbal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TiempoVerbal"
              },
              {
                "type": "null"
              }
            ]
          },
          "titulo": {
            "minLength": 1,
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "titulo",
          "premisa"
        ],
        "title": "BriefEntrada",
        "type": "object"
      },
...
      "CambioPedido": {
        "additionalProperties": false,
        "description": "`SPEC-45` `RF-04`: un cambio del lector, con el estado de su trabajo y el motivo tecnico\nsi fallo. Solo en la administracion: el lector ve un mensaje que se entiende.",
        "properties": {
          "creada_en": {
            "title": "Creada En",
            "type": "string"
          },
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EstadoDeTrabajo"
              },
              {
                "type": "null"
              }
            ]
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "texto": {
            "title": "Texto",
            "type": "string"
          }
        },
        "required": [
          "texto",
          "creada_en",
          "estado",
          "motivo"
        ],
        "title": "CambioPedido",
        "type": "object"
      },
      "CambiosPedidos": {
        "additionalProperties": false,
        "properties": {
          "cambios": {
            "items": {
              "$ref": "#/components/schemas/CambioPedido"
            },
            "title": "Cambios",
            "type": "array"
          }
        },
        "required": [
          "cambios"
        ],
        "title": "CambiosPedidos",
        "type": "object"
      },
...
      "CifrasDeLaMatriz": {
        "additionalProperties": false,
        "properties": {
          "abiertos": {
            "title": "Abiertos",
            "type": "integer"
          },
          "coste": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CosteDeUnaObra"
              },
              {
                "type": "null"
              }
            ]
          },
          "gastado": {
            "$ref": "#/components/schemas/Gastado"
          },
          "techo_usd": {
            "title": "Techo Usd",
            "type": "number"
          }
        },
        "required": [
          "coste",
          "gastado",
          "techo_usd",
          "abiertos"
        ],
        "title": "CifrasDeLaMatriz",
        "type": "object"
      },
...
      "Estanteria": {
        "additionalProperties": false,
        "properties": {
          "obras": {
            "items": {
              "$ref": "#/components/schemas/ObraEnLaEstanteria"
            },
            "title": "Obras",
            "type": "array"
          }
        },
        "required": [
          "obras"
        ],
        "title": "Estanteria",
        "type": "object"
      },
      "EventoDeLaHistoria": {
        "additionalProperties": false,
        "description": "`SPEC-37` `RF-02`: un momento de la historia de una novela. `tipo` es entrevista,\nronda_del_plan, capitulo, parada, ronda_de_la_puerta o version; cada uno rellena lo suyo y\ndeja lo demas vacio. `cuando` es nulo en lo que no tiene hora (las rondas del plan).",
        "properties": {
          "aprobado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aprobado"
          },
          "capitulo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capitulo"
          },
          "capitulos_cambiados": {
            "items": {
              "type": "integer"
            },
            "title": "Capitulos Cambiados",
            "type": "array"
          },
          "codigo_lean": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Lean"
          },
          "condiciones": {
            "items": {
              "type": "string"
            },
            "title": "Condiciones",
            "type": "array"
          },
          "coste": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CosteDeUnaObra"
              },
              {
                "type": "null"
              }
            ]
          },
          "cuando": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuando"
          },
          "escenas": {
            "items": {
              "$ref": "#/components/schemas/EscenaEnGeneracion"
            },
            "title": "Escenas",
            "type": "array"
          },
          "intentos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intentos"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "notas": {
            "items": {
              "$ref": "#/components/schemas/NotaDelEditor"
            },
            "title": "Notas",
            "type": "array"
          },
          "objeciones": {
            "items": {
              "type": "string"
            },
            "title": "Objeciones",
            "type": "array"
          },
          "origen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origen"
          },
          "peticion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peticion"
          },
          "ronda": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ronda"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "tipo",
          "cuando",
          "version",
          "capitulo",
          "coste",
          "notas",
          "escenas",
          "intentos",
          "motivo",
          "aprobado",
          "origen",
          "objeciones",
          "ronda",
          "codigo_lean",
          "condiciones",
          "peticion",
          "capitulos_cambiados"
        ],
        "title": "EventoDeLaHistoria",
        "type": "object"
      },
...
      "GeneracionEnVivo": {
        "additionalProperties": false,
        "description": "`SPEC-33` `RF-14`..`RF-17`. Desde `SPEC-35` `RF-13` (`PLAN-35` F2), tambien la fase\nde la obra -la ultima de su progreso- y el motivo del ultimo lanzamiento si fallo: es lo\nque la pagina dice mientras se planifica, sin capitulos, o si nada llego a empezar.",
        "properties": {
          "capitulos": {
            "items": {
              "$ref": "#/components/schemas/CapituloEnGeneracion"
            },
            "title": "Capitulos",
            "type": "array"
          },
          "coste": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CosteDeLaGeneracion"
              },
              {
                "type": "null"
              }
            ]
          },
          "fase_de_la_obra": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaseDeGeneracion"
              },
              {
                "type": "null"
              }
            ]
          },
          "motivo_del_fallo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Del Fallo"
          },
          "obra": {
            "title": "Obra",
            "type": "string"
          },
          "titulo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Titulo"
          },
          "total_de_capitulos": {
            "title": "Total De Capitulos",
            "type": "integer"
          }
        },
        "required": [
          "obra",
          "total_de_capitulos",
          "capitulos",
          "coste"
        ],
        "title": "GeneracionEnVivo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HallazgoAbiertoDeLaObra": {
        "additionalProperties": false,
        "properties": {
          "capitulo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capitulo"
          },
          "descripcion": {
            "title": "Descripcion",
            "type": "string"
          },
          "invariante": {
            "title": "Invariante",
            "type": "string"
          },
          "severidad": {
            "$ref": "#/components/schemas/Severidad"
          }
        },
        "required": [
          "invariante",
          "severidad",
          "capitulo",
          "descripcion"
        ],
        "title": "HallazgoAbiertoDeLaObra",
        "type": "object"
      },
...
      "HistoriaDeObra": {
        "additionalProperties": false,
        "description": "`SPEC-37`: la historia de una novela, con sus totales al lado. `atribucion` dice como\nse reparte el coste por capitulo (`RF-04`).",
        "properties": {
          "abiertos": {
            "items": {
              "$ref": "#/components/schemas/HallazgoAbiertoDeLaObra"
            },
            "title": "Abiertos",
            "type": "array"
          },
          "atribucion": {
            "title": "Atribucion",
            "type": "string"
          },
          "eventos": {
            "items": {
              "$ref": "#/components/schemas/EventoDeLaHistoria"
            },
            "title": "Eventos",
            "type": "array"
          },
          "obra": {
            "title": "Obra",
            "type": "string"
          },
          "por_agente": {
            "items": {
              "$ref": "#/components/schemas/CosteDeUnAgente"
            },
            "title": "Por Agente",
            "type": "array"
          },
          "titulo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Titulo"
          },
          "totales": {
            "$ref": "#/components/schemas/TotalesDeLaHistoria"
          }
        },
        "required": [
          "obra",
          "titulo",
          "totales",
          "por_agente",
          "abiertos",
          "eventos",
          "atribucion"
        ],
        "title": "HistoriaDeObra",
        "type": "object"
      },
...
      "MatrizDeObra": {
        "additionalProperties": false,
        "description": "`SPEC-38`: la matriz por capitulo de una version de la novela, resuelta en el backend.",
        "properties": {
          "abiertos": {
            "items": {
              "$ref": "#/components/schemas/HallazgoAbiertoDeLaObra"
            },
            "title": "Abiertos",
            "type": "array"
          },
          "atribucion": {
            "title": "Atribucion",
            "type": "string"
          },
          "cifras": {
            "$ref": "#/components/schemas/CifrasDeLaMatriz"
          },
          "filas": {
            "items": {
              "$ref": "#/components/schemas/FilaDeLaMatriz"
            },
            "title": "Filas",
            "type": "array"
          },
          "obra": {
            "title": "Obra",
            "type": "string"
          },
          "paradas": {
            "items": {
              "$ref": "#/components/schemas/EventoDeLaHistoria"
            },
            "title": "Paradas",
            "type": "array"
          },
          "por_agente": {
            "items": {
              "$ref": "#/components/schemas/CosteDeUnAgente"
            },
            "title": "Por Agente",
            "type": "array"
          },
          "puerta": {
            "items": {
              "$ref": "#/components/schemas/EventoDeLaHistoria"
            },
            "title": "Puerta",
            "type": "array"
          },
          "titulo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Titulo"
          },
          "totales": {
            "$ref": "#/components/schemas/TotalesDeLaMatriz"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "versiones": {
            "items": {
              "$ref": "#/components/schemas/VersionDeLaMatriz"
            },
            "title": "Versiones",
            "type": "array"
          }
        },
        "required": [
          "obra",
          "titulo",
          "version",
          "versiones",
          "cifras",
          "filas",
          "totales",
          "paradas",
          "puerta",
          "por_agente",
          "abiertos",
          "atribucion"
        ],
        "title": "MatrizDeObra",
        "type": "object"
      },
...
      "ObraSalida": {
        "additionalProperties": false,
        "description": "Una obra nunca se devuelve sin su estructura.\n\n`RF-23` lo dice para la escena -texto, estado y hallazgos abiertos, nunca el\ntexto solo- y el motivo vale igual aqui: un dato suelto induce a darlo por\nbueno. Una obra recien creada devuelve `capitulos: []`, que es informacion,\nno un hueco.",
        "properties": {
          "capitulos": {
            "items": {
              "type": "string"
            },
            "title": "Capitulos",
            "type": "array"
          },
          "genero": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genero"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "premisa": {
            "title": "Premisa",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "titulo",
          "premisa"
        ],
        "title": "ObraSalida",
        "type": "object"
      },
...
      "PeticionDeLaVersion": {
        "additionalProperties": false,
        "description": "`SPEC-35` `RF-10`: las palabras del lector que originaron la version, o nulo si no\nnacio de una peticion.",
        "properties": {
          "texto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texto"
          }
        },
        "required": [
          "texto"
        ],
        "title": "PeticionDeLaVersion",
        "type": "object"
      },
...
      "RetiradaEntrada": {
        "additionalProperties": false,
        "properties": {
          "motivo": {
            "minLength": 1,
            "title": "Motivo",
            "type": "string"
          },
          "quien": {
            "default": "sistema",
            "minLength": 1,
            "title": "Quien",
            "type": "string"
          }
        },
        "required": [
          "motivo"
        ],
        "title": "RetiradaEntrada",
        "type": "object"
      },
...
      "TiempoVerbal": {
        "enum": [
          "presente",
          "pasado"
        ],
        "title": "TiempoVerbal",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Harness de novelas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/obras/{id_obra}/cambios": {
      "get": {
        "description": "`SPEC-45` `RF-04`: los cambios pedidos de la novela, con el motivo tecnico si fallaron.",
        "operationId": "cambios_admin_obras__id_obra__cambios_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CambiosPedidos"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cambios",
        "tags": [
          "regalo"
        ]
      }
    },
    "/admin/obras/{id_obra}/historia": {
      "get": {
        "description": "`SPEC-37`: la historia de una novela, en orden, con sus totales. Sin login (`SPEC-36`).",
        "operationId": "historia_admin_obras__id_obra__historia_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoriaDeObra"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Historia",
        "tags": [
          "regalo"
        ]
      }
    },
    "/admin/obras/{id_obra}/matriz": {
      "get": {
        "description": "`SPEC-38`: la matriz por capitulo de una version (por defecto, la vigente).",
        "operationId": "matriz_admin_obras__id_obra__matriz_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatrizDeObra"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Matriz",
        "tags": [
          "regalo"
        ]
      }
    },
    "/admin/obras/{id_obra}/retirada": {
      "delete": {
        "description": "`SPEC-41` `RF-02`: devolverla a la estanteria.",
        "operationId": "devolver_admin_obras__id_obra__retirada_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Devolver",
        "tags": [
          "regalo"
        ]
      },
      "post": {
        "description": "`SPEC-41` `RF-01`: quitarla de la estanteria con su motivo. No borra nada.",
        "operationId": "retirar_admin_obras__id_obra__retirada_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetiradaEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retirar",
        "tags": [
          "regalo"
        ]
      }
    },
...
    "/obras": {
      "get": {
        "description": "`RF-01`..`RF-03`: todas las obras, con su portada, su destinatario y su estado.",
        "operationId": "estanteria_obras_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estanteria"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Estanteria",
        "tags": [
          "regalo"
        ]
      },
      "post": {
        "operationId": "alta_de_obra_obras_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObraSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alta De Obra",
        "tags": [
          "obras"
        ]
      }
    },
...
    "/obras/{id_obra}/generacion": {
      "get": {
        "description": "`RF-14`..`RF-17`, `RF-20`: los capitulos con su fase, las notas del Editor al\ncerrarse cada uno y el coste de la ultima generacion.",
        "operationId": "generacion_obras__id_obra__generacion_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneracionEnVivo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generacion",
        "tags": [
          "regalo"
        ]
      }
    },
...
    "/obras/{id_obra}/versiones/{numero}/peticion": {
      "get": {
        "description": "`SPEC-35` `RF-10`: que peticion cambio esta version, con las palabras del lector.",
        "operationId": "peticion_de_la_version_obras__id_obra__versiones__numero__peticion_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_obra",
            "required": true,
            "schema": {
              "title": "Id Obra",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "numero",
            "required": true,
            "schema": {
              "title": "Numero",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeticionDeLaVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Peticion De La Version",
        "tags": [
          "regalo"
        ]
      }
    },
...
  }
}